pyproject.toml
[tool.poetry]
name = "docs-chatbot"
version = "0.1.0"
description = ""
authors = ["afzalfaizi <afzaalm993@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
langchain-google-genai = "^2.0.7"
langchain-core = "^0.3.25"
langgraph = "^0.2.59"
dependencies = "^7.7.1"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
langchain = "^0.3.13"
faiss-cpu = "^1.9.0.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
